+++ docs/source/changelog.rst
@@ -13,6 +13,7 @@ v26.1.12-dev
 - Added :class:`~discord.events.MessageReactionRemoveEvent` event.
 - Added :class:`~discord.enums.ReactionType` enum.
 - Added :class:`~discord.events.MessageReactionAddEvent` event.
+- Added :class:`~discord.events.MessageDeleteBulkEvent` event.
 
 
 v26.1.11